added tree submodules, perf macro, helper funcs

+++ src/zrub/serializer/int32.c
@@ -41,13 +41,13 @@ uint8_t zrub_deserialize_signed_int32(
     // get the integer in big endian (uint32_t)
     *value = 0;
     
-    *value |= buf[*offset] << 24;
+    *value |= ((int32_t)buf[*offset]) << 24;
     (*offset)++;
-    *value |= buf[*offset] << 16;
+    *value |= ((int32_t)buf[*offset]) << 16;
     (*offset)++;
-    *value |= buf[*offset] << 8;
+    *value |= ((int32_t)buf[*offset]) << 8;
     (*offset)++;
-    *value |= buf[*offset];
+    *value |= ((int32_t)buf[*offset]);
     (*offset)++;
 
     return 0;
